refactor(backend): complete Codex catalog ownership
