+++ src-tauri/src/codex_config.rs
@@ -1,18 +1,17 @@
-use std::collections::HashSet;
 use std::path::{Path, PathBuf};
 
-use crate::config::{
-    get_home_dir, path_is_within, read_json_file, write_json_file, write_text_file,
-};
+use crate::config::{read_json_file, write_text_file};
 use crate::error::AppError;
-use crate::model_capabilities::{image_input_capability_from_modalities, ImageInputCapability};
 use crate::provider::{Provider, ProviderMeta};
-use once_cell::sync::OnceCell;
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 use std::fs;
-use std::process::{Command, Stdio};
 use toml_edit::{DocumentMut, Item, TableLike};
+
+#[cfg(test)]
+use once_cell::sync::OnceCell;
+#[cfg(test)]
+use std::collections::HashSet;
 
 mod auth;
 mod catalog;
@@ -30,9 +29,15 @@ use auth::{
 };
 
 #[cfg(test)]
-use catalog::{build_simplified_catalog_from_texts, read_limited_string, MAX_CODEX_CATALOG_BYTES};
-pub use catalog::{read_codex_model_catalog_simplified_from_live, CodexCatalogToolProfile};
-pub(crate) use catalog::{read_codex_model_catalog_text, resolve_fyagent_catalog_path};
+use catalog::*;
+pub(crate) use catalog::{
+    codex_top_level_model, read_codex_model_catalog_text, resolve_fyagent_catalog_path,
+    CODEX_WEB_SEARCH_DISABLED, CODEX_WEB_SEARCH_FIELD,
+};
+pub use catalog::{
+    prepare_codex_config_text_with_model_catalog, read_codex_model_catalog_simplified_from_live,
+    CodexCatalogToolProfile,
+};
 
 #[cfg(test)]
 pub use features::CODEX_WEBSOCKET_NON_GPT_MODEL_WARNING;
@@ -67,98 +72,6 @@ pub const FYAGENT_CODEX_MODEL_PROVIDER_ID: &str = "custom";
 pub const FYAGENT_CODEX_OFFICIAL_PROXY_PROVIDER_ID: &str = "fyagent-official";
 pub const FYAGENT_CODEX_MODEL_CATALOG_FILENAME: &str = "fyagent-model-catalog.json";
 const CODEX_PROXY_AUTH_PLACEHOLDER: &str = "PROXY_MANAGED";
-
-#[cfg(target_os = "windows")]
-const CREATE_NO_WINDOW: u32 = 0x08000000;
-
-// Generating a ProxyChat catalog only needs one stable Codex model template per
-// process. Without this cache every provider switch/takeover can start the
-// Codex CLI again, which is especially expensive for npm-installed `codex.cmd`
-// on Windows. Tests deliberately bypass the global cache because they isolate
-// CODEX_HOME and seed different model templates.
-#[cfg(not(test))]
-static CODEX_MODEL_CATALOG_TEMPLATE_CACHE: OnceCell<Value> = OnceCell::new();
-
-/// Top-level `config.toml` key that controls Codex's built-in web-search tool.
-pub(crate) const CODEX_WEB_SEARCH_FIELD: &str = "web_search";
-/// Value that disables the web-search tool. Some native `/responses` gateways
-/// reject a `web_search` tool with `responses_feature_not_supported` ("tool type
-/// 'web_search' is not supported by this gateway phase"), so for those we write
-/// this per the vendors' official Codex docs. Also doubles as fyagent's
-/// ownership sentinel: we only ever remove a `web_search` key whose value equals
-/// this string, never a user's own setting.
-pub(crate) const CODEX_WEB_SEARCH_DISABLED: &str = "disabled";
-
-/// Native `/responses` gateways whose first-party models do NOT support the Codex
-/// `web_search` hosted tool. A BLACKLIST (default-on): everything not listed keeps
-/// Codex's default, so relays/aggregators fronting real GPT — and any unknown
-/// provider — are never touched. This avoids a whitelist's dangerous failure mode
-/// (a fragile "is this GPT?" heuristic wrongly keeping web_search ON → hard 400);
-/// the blacklist's failure mode is the safe, recoverable one (a not-yet-listed
-/// broken gateway errors once → add it here).
-///
-/// Matched two ways so an aggregator (e.g. SiliconFlow) fronting these vendors'
-/// models is also caught:
-/// - `base_url` host substring, and
-/// - the model id's brand prefix (after stripping any `vendor/` path segment).
-///
-/// Verified 2026-06-28 doc audit — reject: MiMo (hard 400), LongCat (official
-/// config ships `web_search = "disabled"`), MiniMax (tool-type enum `['function']`
-/// only), and Qwen3-Coder models (百炼 marks built-in tools unsupported for
-/// the coder series). Deliberately NOT listed by host: 火山方舟豆包, general
-/// 阿里百炼 Qwen models that support built-in web_search, and GPT-native relays.
-const CODEX_WEB_SEARCH_REJECT_HOSTS: &[&str] = &[
-    "xiaomimimo.com", // Xiaomi MiMo (api.xiaomimimo.com, token-plan-cn.xiaomimimo.com)
-    "longcat.chat",   // Meituan LongCat (api.longcat.chat)
-    "minimax.io",     // MiniMax global (api.minimax.io)
-    "minimaxi.com",   // MiniMax CN (api.minimaxi.com)
-];
-
-/// Brand prefixes of models whose native gateways reject `web_search`, matched
-/// against the model id's last `/`-segment so aggregator ids like
-/// `MiniMaxAI/MiniMax-M3` are caught. Exact brand names (not a fuzzy heuristic),
-/// so a supporting gateway is never wrongly matched.
-const CODEX_WEB_SEARCH_REJECT_MODEL_PREFIXES: &[&str] =
-    &["mimo", "longcat", "minimax", "qwen3-coder"];
-
-/// Top-level `model` id from a Codex `config.toml`.
-pub(crate) fn codex_top_level_model(config_text: &str) -> Option<String> {
-    let doc = config_text.parse::<toml::Value>().ok()?;
-    doc.get("model")
-        .and_then(|value| value.as_str())
-        .map(|value| value.trim().to_string())
-        .filter(|value| !value.is_empty())
-}
-
-/// Whether a native `/responses` provider's gateway is known to reject the Codex
-/// `web_search` hosted tool — by `base_url` host OR by the active model's brand
-/// (so an aggregator fronting a reject vendor's model is caught too). Driven by
-/// the live `config.toml`, so it applies to existing providers without a re-save.
-fn codex_native_gateway_rejects_web_search(config_text: &str) -> bool {
-    if let Some(base_url) = extract_codex_base_url(config_text) {
-        let base_url = base_url.to_ascii_lowercase();
-        if CODEX_WEB_SEARCH_REJECT_HOSTS
-            .iter()
-            .any(|host| base_url.contains(host))
-        {
-            return true;
-        }
-    }
-    if let Some(model) = codex_top_level_model(config_text) {
-        let model = model.to_ascii_lowercase();
-        // Strip any aggregator "vendor/" prefix, e.g. "MiniMaxAI/MiniMax-M3"
-        // or "qwen/qwen3-coder-plus".
-        let model = model.rsplit('/').next().unwrap_or(model.as_str());
-        if CODEX_WEB_SEARCH_REJECT_MODEL_PREFIXES
-            .iter()
-            .any(|prefix| model.starts_with(prefix))
-        {
-            return true;
-        }
-    }
-    false
-}
-const CODEX_MODEL_CATALOG_TEMPLATE_SLUG: &str = "gpt-5.5";
 
 /// Reserved built-in provider IDs from OpenAI Codex's config/model-provider
 /// catalog. Keep in sync with Codex `RESERVED_MODEL_PROVIDER_IDS` and legacy
@@ -236,868 +149,6 @@ pub fn extract_codex_base_url(config_text: &str) -> Option<String> {
     doc.get("base_url")
         .and_then(|v| v.as_str())
         .map(ToString::to_string)
-}
-
-fn parse_codex_positive_u64(value: Option<&Value>) -> Option<u64> {
-    match value {
-        Some(Value::Number(n)) => n.as_u64().filter(|v| *v > 0),
-        Some(Value::String(s)) => s.trim().parse::<u64>().ok().filter(|v| *v > 0),
-        _ => None,
-    }
-}
-
-fn extract_codex_top_level_u64(config_text: &str, field: &str) -> Option<u64> {
-    let doc = config_text.parse::<toml::Value>().ok()?;
-    doc.get(field)
-        .and_then(|value| value.as_integer())
-        .and_then(|value| u64::try_from(value).ok())
-        .filter(|value| *value > 0)
-}
-
-fn codex_catalog_input_modalities(
-    model: &str,
-    declared_modalities: Option<&[String]>,
-) -> Vec<String> {
-    let modalities = match image_input_capability_from_modalities(model, declared_modalities) {
-        ImageInputCapability::Unsupported => &["text"][..],
-        ImageInputCapability::Supported | ImageInputCapability::Unknown => &["text", "image"][..],
-    };
-    modalities.iter().map(|item| (*item).to_string()).collect()
-}
-
-fn codex_catalog_model_entry(
-    template: &Value,
-    spec: &CodexCatalogModelSpec,
-    priority: usize,
-    profile: CodexCatalogToolProfile,
-    default_context_window: u64,
-) -> Value {
-    let mut entry = template.clone();
-    let Some(entry_obj) = entry.as_object_mut() else {
-        return json!({});
-    };
-
-    let display_name = spec.display_name.as_deref().unwrap_or(&spec.model);
-    let context_window = spec.context_window.unwrap_or(default_context_window);
-    entry_obj.insert("slug".to_string(), json!(spec.model));
-    entry_obj.insert("display_name".to_string(), json!(display_name));
-    entry_obj.insert("description".to_string(), json!(display_name));
-    entry_obj.insert("context_window".to_string(), json!(context_window));
-    entry_obj.insert("max_context_window".to_string(), json!(context_window));
-    entry_obj.insert("priority".to_string(), json!(1000 + priority));
-    entry_obj.insert("additional_speed_tiers".to_string(), json!([]));
-    entry_obj.insert("service_tiers".to_string(), json!([]));
-    entry_obj.insert("availability_nux".to_string(), Value::Null);
-    entry_obj.insert("upgrade".to_string(), Value::Null);
-
-    // Image support is a model capability, not a tool-profile capability.
-    // Trust hidden preset metadata first, then the confirmed text-only registry;
-    // every unknown model fails open so GPT/relay aliases are never declared
-    // text-only merely because a template had a conservative default.
-    entry_obj.insert(
-        "input_modalities".to_string(),
-        json!(codex_catalog_input_modalities(
-            &spec.model,
-            spec.input_modalities.as_deref(),
-        )),
-    );
-
-    if profile != CodexCatalogToolProfile::ProxyChat {
-        // Native `/responses` and Anthropic gateways reject / drop Codex's freeform
-        // `apply_patch` (type=="custom") tool. Strip any key that would make Codex
-        // emit a custom/freeform tool, and rely on shell_type="shell_command" for
-        // edits. Defensive even though the native template is already clean
-        // (guards against template drift / an accidental gpt-5.5 clone).
-        //
-        // NOTE: `base_instructions` is NOT stripped — Codex's catalog parser
-        // treats it as a REQUIRED field and refuses to load the file without
-        // it ("missing field `base_instructions`"). The template carries a
-        // neutral identity default; per-vendor official text overrides below.
-        for key in [
-            "apply_patch_tool_type",
-            "web_search_tool_type",
-            "tools",
-            "model_messages",
-        ] {
-            entry_obj.remove(key);
-        }
-        entry_obj.insert("shell_type".to_string(), json!("shell_command"));
-
-        if let Some(base_instructions) = spec
-            .base_instructions
-            .as_deref()
-            .map(str::trim)
-            .filter(|s| !s.is_empty())
-        {
-            entry_obj.insert("base_instructions".to_string(), json!(base_instructions));
-        }
-        if let Some(parallel) = spec.supports_parallel_tool_calls {
-            entry_obj.insert("supports_parallel_tool_calls".to_string(), json!(parallel));
-        }
-    }
-
-    entry
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct CodexCatalogModelSpec {
-    model: String,
-    /// Explicit user value only. Entries fall back to the model id — except
-    /// official vendor catalog entries, which keep the vendor's display name.
-    display_name: Option<String>,
-    /// Explicit user value only. Entries fall back to the config's
-    /// `model_context_window` (or 128k) — except official vendor catalog
-    /// entries, which keep the vendor's declared window.
-    context_window: Option<u64>,
-    /// Per-row override for the native template's `supports_parallel_tool_calls`
-    /// (e.g. MiniMax=true, MiMo=false). Only consulted for `NativeResponses`.
-    supports_parallel_tool_calls: Option<bool>,
-    /// Hidden per-row capability declaration from built-in provider metadata.
-    /// When omitted, all catalog profiles consult the shared text-only model
-    /// registry and otherwise default to `["text", "image"]`.
-    input_modalities: Option<Vec<String>>,
-    /// Per-row override for the native template's `base_instructions` (the
-    /// model identity / system preamble). Carries each vendor's OFFICIAL value
-    /// (e.g. MiMo "developed by Xiaomi", MiniMax "based on MiniMax-M3"); falls
-    /// back to the template default when absent. Only consulted for
-    /// `NativeResponses`.
-    base_instructions: Option<String>,
-}
-
-fn codex_catalog_model_specs(settings: &Value) -> Vec<CodexCatalogModelSpec> {
-    let Some(models) = settings
-        .get("modelCatalog")
-        .and_then(|catalog| catalog.get("models"))
-        .and_then(|models| models.as_array())
-    else {
-        return Vec::new();
-    };
-
-    let mut seen = std::collections::HashSet::new();
-    let mut specs = Vec::new();
-
-    for model_config in models {
-        let Some(model) = model_config
-            .get("model")
-            .and_then(|value| value.as_str())
-            .map(str::trim)
-            .filter(|model| !model.is_empty())
-        else {
-            continue;
-        };
-
-        if !seen.insert(model.to_string()) {
-            continue;
-        }
-
-        let display_name = model_config
-            .get("displayName")
-            .or_else(|| model_config.get("display_name"))
-            .and_then(|value| value.as_str())
-            .map(str::trim)
-            .filter(|name| !name.is_empty())
-            .map(str::to_string);
-        let context_window = parse_codex_positive_u64(
-            model_config
-                .get("contextWindow")
-                .or_else(|| model_config.get("context_window")),
-        );
-
-        let supports_parallel_tool_calls = model_config
-            .get("supportsParallelToolCalls")
-            .or_else(|| model_config.get("supports_parallel_tool_calls"))
-            .and_then(|value| value.as_bool());
-        let input_modalities = model_config
-            .get("inputModalities")
-            .or_else(|| model_config.get("input_modalities"))
-            .and_then(|value| value.as_array())
-            .map(|items| {
-                items
-                    .iter()
-                    .filter_map(|item| item.as_str())
-                    .map(str::to_string)
-                    .collect::<Vec<_>>()
-            })
-            .filter(|items| !items.is_empty());
-
-        let base_instructions = model_config
-            .get("baseInstructions")
-            .or_else(|| model_config.get("base_instructions"))
-            .and_then(|value| value.as_str())
-            .map(str::trim)
-            .filter(|text| !text.is_empty())
-            .map(str::to_string);
-
-        specs.push(CodexCatalogModelSpec {
-            model: model.to_string(),
-            display_name,
-            context_window,
-            supports_parallel_tool_calls,
-            input_modalities,
-            base_instructions,
-        });
-    }
-
-    specs
-}
-
-fn find_codex_model_template(catalog: &Value) -> Option<Value> {
-    catalog
-        .get("models")
-        .and_then(|models| models.as_array())
-        .and_then(|models| {
-            models.iter().find(|model| {
-                model.get("slug").and_then(|slug| slug.as_str())
-                    == Some(CODEX_MODEL_CATALOG_TEMPLATE_SLUG)
-            })
-        })
-        .cloned()
-}
-
-fn load_codex_model_template_from_cache() -> Result<Option<Value>, AppError> {
-    let path = get_codex_config_dir().join("models_cache.json");
-    if !path.exists() {
-        return Ok(None);
-    }
-
-    let text = fs::read_to_string(&path).map_err(|e| AppError::io(&path, e))?;
-    let catalog: Value = serde_json::from_str(&text).map_err(|e| AppError::json(&path, e))?;
-    Ok(find_codex_model_template(&catalog))
-}
-
-/// Fixed candidates for locating the `codex` CLI when it is not on the process
-/// PATH (common in GUI apps launched outside a terminal).
-#[cfg(target_os = "macos")]
-const CODEX_CLI_FIXED_CANDIDATES: &[&str] = &[
-    "codex",                   // macOS PATH
-    "/opt/homebrew/bin/codex", // macOS Apple Silicon Homebrew
-    "/usr/local/bin/codex",    // macOS Intel Homebrew
-];
-
-#[cfg(windows)]
-const CODEX_CLI_FIXED_CANDIDATES: &[&str] = &[];
-
-fn push_codex_cli_candidate(
-    candidates: &mut Vec<PathBuf>,
-    seen: &mut HashSet<String>,
-    candidate: PathBuf,
-) {
-    let key = candidate.to_string_lossy().into_owned();
-    if seen.insert(key) {
-        candidates.push(candidate);
-    }
-}
-
-fn push_existing_codex_cli_candidate(
-    candidates: &mut Vec<PathBuf>,
-    seen: &mut HashSet<String>,
-    candidate: PathBuf,
-) {
-    if candidate.exists() {
-        push_codex_cli_candidate(candidates, seen, candidate);
-    }
-}
-
-fn push_codex_cli_candidates_from_version_dirs(
-    candidates: &mut Vec<PathBuf>,
-    seen: &mut HashSet<String>,
-    versions_dir: PathBuf,
-    suffix: &[&str],
-) {
-    let Ok(entries) = fs::read_dir(versions_dir) else {
-        return;
-    };
-
-    let mut discovered = entries
-        .filter_map(Result::ok)
-        .map(|entry| {
-            let mut candidate = entry.path();
-            for component in suffix {
-                candidate.push(component);
-            }
-            candidate
-        })
-        .filter(|candidate| candidate.exists())
-        .collect::<Vec<_>>();
-
-    // Prefer newer-looking version directories before older global installs.
-    discovered.sort_by(|a, b| b.cmp(a));
-    for candidate in discovered {
-        push_codex_cli_candidate(candidates, seen, candidate);
-    }
-}
-
-fn push_home_codex_cli_candidates(
-    candidates: &mut Vec<PathBuf>,
-    seen: &mut HashSet<String>,
-    home: &Path,
-) {
-    for relative in [
-        ".nvm/current/bin/codex",
-        ".volta/bin/codex",
-        ".asdf/shims/codex",
-        ".local/share/mise/shims/codex",
-        ".config/mise/shims/codex",
-        ".local/bin/codex",
-        ".npm-global/bin/codex",
-        ".npm-packages/bin/codex",
-        ".local/share/pnpm/codex",
-        "Library/pnpm/codex",
-    ] {
-        push_existing_codex_cli_candidate(candidates, seen, home.join(relative));
-    }
-
-    push_codex_cli_candidates_from_version_dirs(
-        candidates,
-        seen,
-        home.join(".nvm/versions/node"),
-        &["bin", "codex"],
-    );
-    push_codex_cli_candidates_from_version_dirs(
-        candidates,
-        seen,
-        home.join(".local/share/fnm/node-versions"),
-        &["installation", "bin", "codex"],
-    );
-    push_codex_cli_candidates_from_version_dirs(
-        candidates,
-        seen,
-        home.join("Library/Application Support/fnm/node-versions"),
-        &["installation", "bin", "codex"],
-    );
-}
-
-#[cfg(target_os = "macos")]
-fn push_env_codex_cli_candidates(candidates: &mut Vec<PathBuf>, seen: &mut HashSet<String>) {
-    for (env_key, suffix) in [
-        ("NPM_CONFIG_PREFIX", &["bin", "codex"][..]),
-        ("VOLTA_HOME", &["bin", "codex"][..]),
-        ("ASDF_DATA_DIR", &["shims", "codex"][..]),
-        ("MISE_DATA_DIR", &["shims", "codex"][..]),
-        ("PNPM_HOME", &["codex"][..]),
-    ] {
-        let Some(prefix) = std::env::var_os(env_key) else {
-            continue;
-        };
-        let mut candidate = PathBuf::from(prefix);
-        for component in suffix {
-            candidate.push(component);
-        }
-        push_existing_codex_cli_candidate(candidates, seen, candidate);
-    }
-
-    if let Some(nvm_dir) = std::env::var_os("NVM_DIR") {
-        push_codex_cli_candidates_from_version_dirs(
-            candidates,
-            seen,
-            PathBuf::from(nvm_dir).join("versions/node"),
-            &["bin", "codex"],
-        );
-    }
-
-    if let Some(fnm_dir) = std::env::var_os("FNM_DIR") {
-        push_codex_cli_candidates_from_version_dirs(
-            candidates,
-            seen,
-            PathBuf::from(fnm_dir).join("node-versions"),
-            &["installation", "bin", "codex"],
-        );
-    }
-}
-
-#[cfg(windows)]
-fn push_env_codex_cli_candidates(candidates: &mut Vec<PathBuf>, seen: &mut HashSet<String>) {
-    // Never consume elevated-process user tool variables on Windows. The
-    // ordinary Codex CLI candidates come only from Alice's frozen directories
-    // and OS-resolved system locations.
-    for directory in crate::windows_runtime::safe_command_search_paths() {
-        for name in ["codex.cmd", "codex.exe", "codex"] {
-            push_existing_codex_cli_candidate(candidates, seen, directory.join(name));
-        }
-    }
-}
-
-fn codex_cli_candidates() -> Vec<PathBuf> {
-    let mut candidates = Vec::new();
-    let mut seen = HashSet::new();
-
-    for candidate in CODEX_CLI_FIXED_CANDIDATES {
-        push_codex_cli_candidate(&mut candidates, &mut seen, PathBuf::from(candidate));
-    }
-
-    push_env_codex_cli_candidates(&mut candidates, &mut seen);
-    let home = get_home_dir();
-    #[cfg(windows)]
-    if crate::windows_runtime::is_local_command_path(&home) {
-        push_home_codex_cli_candidates(&mut candidates, &mut seen, &home);
-    }
-    #[cfg(target_os = "macos")]
-    push_home_codex_cli_candidates(&mut candidates, &mut seen, &home);
-
-    candidates
-}
-
-fn codex_bundled_cli_allowed(target_is_windows: bool, formal_windows_build: bool) -> bool {
-    !target_is_windows || !formal_windows_build
-}
-
-fn codex_bundled_models_command(
-    candidate: &Path,
-) -> Result<Command, crate::windows_runtime::WindowsStartupErrorCode> {
-    let mut command = Command::new(candidate);
-    command
-        .args(["debug", "models", "--bundled"])
-        .stdin(Stdio::null());
-
-    // A release build uses the Windows GUI subsystem, so a console child that
-    // is created without this flag gets its own transient console window. npm
-    // installs Codex as `codex.cmd`, which Windows launches through cmd.exe.
-    #[cfg(target_os = "windows")]
-    {
-        use std::os::windows::process::CommandExt;
-        crate::windows_runtime::configure_shell_user_command(&mut command, candidate.parent())?;
-        command
-            .current_dir(crate::windows_runtime::require_interactive_user_context().user_profile());
-        command.creation_flags(CREATE_NO_WINDOW);
-    }
-
-    Ok(command)
-}
-
-fn load_codex_model_template_from_bundled() -> Result<Option<Value>, AppError> {
-    if !codex_bundled_cli_allowed(
-        cfg!(target_os = "windows"),
-        crate::windows_runtime::formal_windows_build(),
-    ) {
-        return Ok(None);
-    }
-
-    for candidate in codex_cli_candidates() {
-        let candidate_label = candidate.to_string_lossy();
-        let mut command = match codex_bundled_models_command(&candidate) {
-            Ok(command) => command,
-            Err(error) => {
-                log::debug!("failed to configure Codex CLI environment: {error}");
-                continue;
-            }
-        };
-        let output = match command.output() {
-            Ok(output) => output,
-            Err(err) => {
-                log::debug!("failed to run `{candidate_label} debug models --bundled`: {err}");
-                continue;
-            }
-        };
-
-        if !output.status.success() {
-            let stderr = String::from_utf8_lossy(&output.stderr);
-            log::debug!("`{candidate_label} debug models --bundled` failed: {stderr}");
-            continue;
-        }
-
-        let catalog: Value = match serde_json::from_slice(&output.stdout) {
-            Ok(catalog) => catalog,
-            Err(e) => {
-                log::debug!(
-                    "Failed to parse `{candidate_label} debug models --bundled` output: {e}"
-                );
-                continue;
-            }
-        };
-        if let Some(template) = find_codex_model_template(&catalog) {
-            return Ok(Some(template));
-        }
-    }
-
-    Ok(None)
-}
-
-fn load_codex_model_template_static() -> Option<Value> {
-    let text = include_str!("resources/gpt5_5_template.json");
-    match serde_json::from_str(text) {
-        Ok(template) => Some(template),
-        Err(e) => {
-            log::warn!("Failed to parse bundled gpt-5.5 template: {e}");
-            None
-        }
-    }
-}
-
-/// Bundled clean template for native `/responses` providers. Unlike the
-/// gpt-5.5 template it carries NO freeform `apply_patch` / `web_search` tool
-/// declarations and no GPT-5 base_instructions, so Codex never emits a
-/// `type=="custom"` tool that native gateways (MiMo/MiniMax/…) reject. Edits
-/// flow through `shell_type="shell_command"` instead. We deliberately do NOT
-/// fall back to `models_cache.json` here (that would reintroduce gpt-5.5's
-/// freeform apply_patch).
-fn load_codex_native_responses_template() -> Value {
-    let text = include_str!("resources/codex_native_responses_template.json");
-    serde_json::from_str(text).expect("bundled codex native responses template must be valid JSON")
-}
-
-/// Hosts whose native `/responses` gateway publishes an OFFICIAL Codex model
-/// catalog (models.json) that FyAgent mirrors verbatim. Matched against
-/// `base_url` ONLY — deliberately NOT by model brand, unlike
-/// `CODEX_WEB_SEARCH_REJECT_MODEL_PREFIXES`: the official entries GRANT
-/// capabilities (freeform `apply_patch`, vendor harness), and an aggregator
-/// merely hosting the same model may not honor them. The safe failure
-/// direction for aggregators is the neutral template (degraded but working);
-/// wrongly granting freeform apply_patch would reintroduce the custom-tool
-/// rejection bug.
-const CODEX_DEEPSEEK_OFFICIAL_CATALOG_HOSTS: &[&str] = &["deepseek.com"];
-
-/// Bundled copy of DeepSeek's official Codex models.json — the exact file
-/// their one-click integration script writes (api-docs.deepseek.com →
-/// quick_start/agent_integrations/codex): freeform apply_patch, GPT-5 harness
-/// base_instructions, low/high/max reasoning levels, web_search supported,
-/// 1m context. Declares `minimal_client_version` 0.144.0.
-fn load_codex_deepseek_official_catalog_models() -> Vec<Value> {
-    let text = include_str!("resources/codex_deepseek_catalog_template.json");
-    let catalog: Value =
-        serde_json::from_str(text).expect("bundled DeepSeek official catalog must be valid JSON");
-    catalog
-        .get("models")
-        .and_then(|models| models.as_array())
-        .cloned()
-        .unwrap_or_default()
-}
-
-/// Official vendor catalog entries for the provider in `config_text`, if its
-/// gateway ships one. Only the `NativeResponses` profile qualifies: ProxyChat
-/// runs through FyAgent's converter (gpt-5.5 template contract) and the
-/// Anthropic transform drops custom tools, so both must keep their existing
-/// templates. Host-driven like the web_search blacklist, so existing providers
-/// pick it up on their next switch without a re-save.
-fn codex_official_vendor_catalog_models(
-    config_text: &str,
-    profile: CodexCatalogToolProfile,
-) -> Option<Vec<Value>> {
-    if profile != CodexCatalogToolProfile::NativeResponses {
-        return None;
-    }
-    let base_url = extract_codex_base_url(config_text)?;
-    let parsed_base_url = url::Url::parse(&base_url).ok()?;
-    if parsed_base_url.scheme() != "https" {
-        return None;
-    }
-    let base_url_host = parsed_base_url
-        .host_str()?
-        .trim_end_matches('.')
-        .to_ascii_lowercase();
-    if CODEX_DEEPSEEK_OFFICIAL_CATALOG_HOSTS
-        .iter()
-        .any(|official_host| {
-            base_url_host == *official_host
-                || base_url_host
-                    .strip_suffix(*official_host)
-                    .is_some_and(|prefix| prefix.ends_with('.'))
-        })
-    {
-        let models = load_codex_deepseek_official_catalog_models();
-        if !models.is_empty() {
-            return Some(models);
-        }
-    }
-    None
-}
-
-/// Build one catalog entry from an official vendor catalog: match the user's
-/// model id against the vendor entries by slug; an unknown id clones the
-/// vendor's first (flagship) entry so it keeps the gateway's capability
-/// profile without impersonating the flagship. The official entry is
-/// authoritative — no tool-profile stripping — but explicit per-row user
-/// overrides still win.
-fn codex_vendor_catalog_model_entry(
-    vendor_models: &[Value],
-    spec: &CodexCatalogModelSpec,
-    priority: usize,
-) -> Value {
-    let matched = vendor_models.iter().find(|entry| {
-        entry
-            .get("slug")
-            .and_then(|slug| slug.as_str())
-            .is_some_and(|slug| slug.eq_ignore_ascii_case(&spec.model))
-    });
-    let mut entry = match matched {
-        Some(found) => found.clone(),
-        None => vendor_models.first().cloned().unwrap_or_else(|| json!({})),
-    };
-    let Some(entry_obj) = entry.as_object_mut() else {
-        return json!({});
-    };
-
-    if matched.is_none() {
-        let display_name = spec.display_name.as_deref().unwrap_or(&spec.model);
-        entry_obj.insert("slug".to_string(), json!(spec.model));
-        entry_obj.insert("display_name".to_string(), json!(display_name));
-        entry_obj.insert("description".to_string(), json!(display_name));
-        entry_obj.insert("priority".to_string(), json!(1000 + priority));
-    }
-
-    // Explicit user overrides win over the official entry; absent values keep
-    // the vendor's declarations (context window, modalities, harness, ...).
-    if let Some(display_name) = spec.display_name.as_deref() {
-        entry_obj.insert("display_name".to_string(), json!(display_name));
-    }
-    if let Some(context_window) = spec.context_window {
-        entry_obj.insert("context_window".to_string(), json!(context_window));
-        entry_obj.insert("max_context_window".to_string(), json!(context_window));
-    }
-    if let Some(parallel) = spec.supports_parallel_tool_calls {
-        entry_obj.insert("supports_parallel_tool_calls".to_string(), json!(parallel));
-    }
-    if let Some(modalities) = spec.input_modalities.as_deref() {
-        entry_obj.insert("input_modalities".to_string(), json!(modalities));
-    }
-    if let Some(base_instructions) = spec
-        .base_instructions
-        .as_deref()
-        .map(str::trim)
-        .filter(|text| !text.is_empty())
-    {
-        entry_obj.insert("base_instructions".to_string(), json!(base_instructions));
-    }
-
-    // Defensive: if a future codex parser requires a field the vendor file
-    // predates, backfill only whitelisted parser-required keys.
-    fill_template_fields_from_static(&mut entry);
-    entry
-}
-
-/// Fields Codex's external-catalog parser REQUIRES (no serde default): when
-/// one is missing Codex rejects the whole catalog file at startup ("missing
-/// field ..."). `base_instructions` is the other known required field; the
-/// templates always carry it and `codex_catalog_model_entry` handles it.
-/// When Codex requires a new field, add it here AND to the static templates.
-const CODEX_CATALOG_PARSER_REQUIRED_FIELDS: &[&str] = &["supports_reasoning_summaries"];
-
-/// `models_cache.json` is shared by every Codex install on the machine (npm
-/// CLI, desktop-bundled binary, ...), and each version serializes its own
-/// `ModelInfo` shape — the cache's field set follows whichever process wrote
-/// it last, so it cannot be assumed to satisfy the current external-catalog
-/// schema (observed live: 0.144.5 requires `supports_reasoning_summaries`
-/// while a coexisting build kept rewriting the cache without it). Backfill
-/// ONLY parser-required fields from the bundled static template: optional
-/// capability fields keep their missing-means-default semantics, and existing
-/// values always win.
-fn fill_template_fields_from_static(template: &mut Value) {
-    let Some(static_template) = load_codex_model_template_static() else {
-        return;
-    };
-    let (Some(template_obj), Some(static_obj)) =
-        (template.as_object_mut(), static_template.as_object())
-    else {
-        return;
-    };
-    for key in CODEX_CATALOG_PARSER_REQUIRED_FIELDS {
-        if !template_obj.contains_key(*key) {
-            if let Some(value) = static_obj.get(*key) {
-                template_obj.insert((*key).to_string(), value.clone());
-            }
-        }
-    }
-}
-
-fn load_codex_model_catalog_template_uncached() -> Result<Value, AppError> {
-    // ① models_cache.json (created by Codex when it connects to OpenAI)
-    if let Some(mut template) = load_codex_model_template_from_cache()? {
-        fill_template_fields_from_static(&mut template);
-        return Ok(template);
-    }
-    // ② codex CLI (PATH + platform-specific common paths)
-    if let Some(mut template) = load_codex_model_template_from_bundled()? {
-        fill_template_fields_from_static(&mut template);
-        return Ok(template);
-    }
-    // ③ Static fallback bundled at compile time
-    if let Some(template) = load_codex_model_template_static() {
-        return Ok(template);
-    }
-
-    Err(AppError::Message(format!(
-        "Codex model catalog template `{CODEX_MODEL_CATALOG_TEMPLATE_SLUG}` not found. Please start Codex once so models_cache.json is available, or ensure the `codex` CLI is on PATH."
-    )))
-}
-
-fn get_or_load_codex_model_catalog_template<F>(
-    cache: &OnceCell<Value>,
-    loader: F,
-) -> Result<Value, AppError>
-where
-    F: FnOnce() -> Result<Value, AppError>,
-{
-    cache.get_or_try_init(loader).cloned()
-}
-
-#[cfg(not(test))]
-fn load_codex_model_catalog_template() -> Result<Value, AppError> {
-    get_or_load_codex_model_catalog_template(
-        &CODEX_MODEL_CATALOG_TEMPLATE_CACHE,
-        load_codex_model_catalog_template_uncached,
-    )
-}
-
-#[cfg(test)]
-fn load_codex_model_catalog_template() -> Result<Value, AppError> {
-    load_codex_model_catalog_template_uncached()
-}
-
-fn codex_model_catalog_from_specs(
-    specs: &[CodexCatalogModelSpec],
-    template: &Value,
-    profile: CodexCatalogToolProfile,
-    default_context_window: u64,
-) -> Value {
-    let entries: Vec<Value> = specs
-        .iter()
-        .enumerate()
-        .map(|(index, spec)| {
-            codex_catalog_model_entry(template, spec, index, profile, default_context_window)
-        })
-        .collect();
-
-    json!({ "models": entries })
-}
-
-fn codex_model_catalog_from_settings(
-    settings: &Value,
-    config_text: &str,
-    profile: CodexCatalogToolProfile,
-) -> Result<Option<Value>, AppError> {
-    let specs = codex_catalog_model_specs(settings);
-    if specs.is_empty() {
-        return Ok(None);
-    }
-
-    // Vendors that publish an OFFICIAL Codex models.json for their native
-    // `/responses` gateway get it mirrored verbatim instead of the neutral
-    // template: its freeform apply_patch, vendor harness base_instructions and
-    // reasoning levels are load-bearing (the harness tells the model to use
-    // apply_patch, so catalog and harness must stay consistent).
-    if let Some(vendor_models) = codex_official_vendor_catalog_models(config_text, profile) {
-        let entries: Vec<Value> = specs
-            .iter()
-            .enumerate()
-            .map(|(index, spec)| codex_vendor_catalog_model_entry(&vendor_models, spec, index))
-            .collect();
-        return Ok(Some(json!({ "models": entries })));
-    }
-
-    let default_context_window =
-        extract_codex_top_level_u64(config_text, "model_context_window").unwrap_or(128_000);
-
-    // Native providers use the bundled clean template (no freeform apply_patch,
-    // no cache dependency); proxy-chat providers keep cloning Codex's gpt-5.5
-    // entry so the proxy can rewrite custom<->function tools as before.
-    let template = match profile {
-        CodexCatalogToolProfile::NativeResponses | CodexCatalogToolProfile::Anthropic => {
-            load_codex_native_responses_template()
-        }
-        CodexCatalogToolProfile::ProxyChat => load_codex_model_catalog_template()?,
-    };
-    Ok(Some(codex_model_catalog_from_specs(
-        &specs,
-        &template,
-        profile,
-        default_context_window,
-    )))
-}
-
-fn set_codex_model_catalog_json_field(
-    config_text: &str,
-    catalog_path: Option<&Path>,
-) -> Result<String, AppError> {
-    let mut doc = config_text
-        .parse::<DocumentMut>()
-        .map_err(|e| AppError::Message(format!("Invalid Codex config.toml: {e}")))?;
-
-    match catalog_path {
-        Some(_) => {
-            doc["model_catalog_json"] = toml_edit::value(FYAGENT_CODEX_MODEL_CATALOG_FILENAME);
-        }
-        None => {
-            let should_remove = doc
-                .get("model_catalog_json")
-                .and_then(|item| item.as_str())
-                .map(|path| {
-                    Path::new(path).file_name().and_then(|name| name.to_str())
-                        == Some(FYAGENT_CODEX_MODEL_CATALOG_FILENAME)
-                })
-                .unwrap_or(false);
-            if should_remove {
-                doc.as_table_mut().remove("model_catalog_json");
-            }
-        }
-    }
-
-    Ok(doc.to_string())
-}
-
-/// Pure toggle for the top-level `web_search` field that turns Codex's built-in
-/// web-search tool off. When `disable` is true we write `web_search = "disabled"`
-/// (the catalog's `supports_search_tool` does NOT gate this — the request-time
-/// tool comes from the config, defaulting on). When false we *remove* the field,
-/// but only when it carries fyagent's own `"disabled"` sentinel, so switching
-/// back to a web-search-capable provider re-enables it without clobbering a
-/// user's manual setting.
-///
-/// The caller decides `disable` (see `codex_native_gateway_rejects_web_search`);
-/// lifecycle is bound to the fyagent catalog pointer so the field is set/cleaned
-/// up wherever the native catalog is written/removed.
-fn set_codex_native_web_search_field(config_text: &str, disable: bool) -> Result<String, AppError> {
-    let mut doc = config_text
-        .parse::<DocumentMut>()
-        .map_err(|e| AppError::Message(format!("Invalid Codex config.toml: {e}")))?;
-
-    if disable {
-        doc[CODEX_WEB_SEARCH_FIELD] = toml_edit::value(CODEX_WEB_SEARCH_DISABLED);
-    } else {
-        let owned = doc
-            .get(CODEX_WEB_SEARCH_FIELD)
-            .and_then(|item| item.as_str())
-            == Some(CODEX_WEB_SEARCH_DISABLED);
-        if owned {
-            doc.as_table_mut().remove(CODEX_WEB_SEARCH_FIELD);
-        }
-    }
-
-    Ok(doc.to_string())
-}
-
-/// Generate Codex `model_catalog_json` from provider settings and inject/remove
-/// the top-level TOML field that points Codex to the generated file.
-pub fn prepare_codex_config_text_with_model_catalog(
-    settings: &Value,
-    config_text: &str,
-    profile: CodexCatalogToolProfile,
-) -> Result<String, AppError> {
-    if let Some(catalog) = codex_model_catalog_from_settings(settings, config_text, profile)? {
-        let catalog_path = get_codex_model_catalog_path();
-        let config_text = set_codex_model_catalog_json_field(config_text, Some(&catalog_path))?;
-        // Disable web_search only for native gateways on the reject blacklist
-        // (MiMo/LongCat/MiniMax by host or model brand; Qwen3-Coder by model).
-        // Everything else — relays, DouBao, web-search-capable Qwen models,
-        // unknown providers — keeps Codex's default.
-        let disable_web_search = match profile {
-            // The Responses→Anthropic transform silently drops the Codex web_search
-            // hosted tool, so always disable it here rather than present a dead tool.
-            CodexCatalogToolProfile::Anthropic => true,
-            CodexCatalogToolProfile::NativeResponses => {
-                codex_native_gateway_rejects_web_search(&config_text)
-            }
-            CodexCatalogToolProfile::ProxyChat => false,
-        };
-        let config_text = set_codex_native_web_search_field(&config_text, disable_web_search)?;
-        write_json_file(&catalog_path, &catalog)?;
-        Ok(config_text)
-    } else {
-        let config_text = set_codex_model_catalog_json_field(config_text, None)?;
-        // Even without a generated catalog, the Responses→Anthropic transform drops the
-        // Codex web_search hosted tool, so keep the invariant that an Anthropic provider
-        // never presents it as a dead tool.
-        let disable_web_search = profile == CodexCatalogToolProfile::Anthropic;
-        set_codex_native_web_search_field(&config_text, disable_web_search)
-    }
 }
 
 /// Decide the `config.toml` text to write during a takeover-off restore,
